+++ adapters/amqp-vertx-base/src/test/java/org/eclipse/hono/adapter/amqp/VertxBasedAmqpProtocolAdapterTest.java
@@ -57,7 +57,7 @@ import org.eclipse.hono.adapter.resourcelimits.ResourceLimitChecks;
 import org.eclipse.hono.adapter.test.ProtocolAdapterTestSupport;
 import org.eclipse.hono.auth.Device;
 import org.eclipse.hono.client.ClientErrorException;
-import org.eclipse.hono.client.amqp.connection.AmqpConstants;
+import org.eclipse.hono.client.amqp.connection.AmqpUtils;
 import org.eclipse.hono.client.command.CommandConsumer;
 import org.eclipse.hono.client.command.CommandContext;
 import org.eclipse.hono.client.command.CommandResponse;
@@ -232,7 +232,7 @@ public class VertxBasedAmqpProtocolAdapterTest extends ProtocolAdapterTestSuppor
 
         // THEN the adapter's open frame contains the ANONYMOUS-RELAY capability
         verify(deviceConnection).setOfferedCapabilities(
-                argThat(caps -> Arrays.asList(caps).contains(AmqpConstants.CAP_ANONYMOUS_RELAY)));
+                argThat(caps -> Arrays.asList(caps).contains(AmqpUtils.CAP_ANONYMOUS_RELAY)));
     }