+++ clients/amqp-connection/src/main/java/org/eclipse/hono/client/amqp/tracing/MessageAnnotationsExtractAdapter.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.tracing;
+package org.eclipse.hono.client.amqp.tracing;
 
 import java.util.AbstractMap;
 import java.util.Collections;